// packages/core/src/__generated__/base.ts
// AUTO-GENERATED - DO NOT EDIT
// Run "pnpm build:core" to regenerate

export const base = [
  {
    "name": "base",
    "type": "registry:style",
    "dependencies": [
      "tailwind-variants",
      "clsx",
      "tailwind-merge",
      "react-aria-components",
      "tailwindcss-react-aria-components",
      "tw-animate-css",
      "tailwindcss-autocontrast"
    ],
    "registryDependencies": [
      "utils",
      "focus-styles",
      "theme"
    ],
    "extends": "none",
    "css": {
      "@plugin tailwindcss-react-aria-components": {},
      "@plugin tailwindcss-autocontrast": {}
    },
    "files": []
  }
] as const;